+++ app/controllers/answer_forms_controller.rb
@@ -6,7 +6,6 @@ class AnswerFormsController < ApplicationController
   def new
     # whereメソッドを実施することにより、配列で、qustionのレコードが格納される。
     @questions = Question.where(user_id: params[:user_id])
-    @question = Question.new
     @answer_form = AnswerForm.new
     @user = User.find_by(id: current_user.id)
   end
@@ -14,15 +13,9 @@ class AnswerFormsController < ApplicationController
   def create
     @user = User.find(current_user.id)
     @correct_answers = Question.where(user_id: current_user.id)
-    @user_answers = user_answer_params
-    @answer_form = AnswerForm.new(@user_answers)
-    @question = Question.new
-    @questions = @user.questions
-    unless @answer_form.valid?
-      render :new, status: :unprocessable_entity
-      return
-    end
-    if @answer_form.correct?(@user_answers, @correct_answers) == 3
+    @answer_form = AnswerForm.new
+    @questions = current_user.questions
+    if @answer_form.correct?(user_answer_params, @correct_answers) == 3
       @got_up = GotUp.new
       @got_up.save(@user)
       redirect_to schedules_index_path, notice: t('.success')
@@ -38,15 +31,6 @@ class AnswerFormsController < ApplicationController
     params.require(:answer_form).permit(:answer0, :answer1, :answer2)
   end
 
-  def login_with_token
-    unless logged_in?
-      if params[:token]
-        token = params[:token]
-        user = User.find_by(auth_token: token)
-        user&.ensure_auth_token
-        auto_login(user)
-      end
-    end
-  end
+
 
 end
