--- app/controllers/line_bot_controller.rb
@@ -1,40 +1,0 @@
-class LineBotController < ApplicationController
-  require 'line/bot'  # gem 'line-bot-api'
-
-  # callbackアクションのCSRFトークン認証を無効
-  protect_from_forgery :except => [:callback]
-
-  def client
-    @client ||= Line::Bot::Client.new { |config|
-      config.channel_secret = ENV["LINE_CHANNEL_SECRET"]
-      config.channel_token = ENV["LINE_CHANNEL_TOKEN"]
-    }
-  end
-  
-  def callback
-    body = request.body.read
-    events = client.parse_events_from(body)
-
-    events.each { |event|
-      case event
-      when Line::Bot::Event::Message
-        case event.type
-        when Line::Bot::Event::MessageType::Text
-          message = {
-            type: 'text',
-            text: event.message['text']
-          }
-          client.reply_message(event['replyToken'], message)
-        when Line::Bot::Event::MessageType::Follow #友達登録イベント
-          userId = event['source']['userId'] 
-          User.find_or_create_by(uid: userId)
-        when Line::Bot::Event::MessageType::Unfollow #友達削除イベント
-          userId = event['source']['userId']  
-          user = User.find_by(uid: userId)
-          user.destroy if user.present?
-        end
-      end
-    }
-    "OK"
-  end
-end
