+++ app/views/answer_forms/new.html.erb
@@ -1,7 +1,7 @@
 <div class="flex justify-center bg-gray-100">
   <div class="w-1/3 h-2/3 shadow-lg rounded-3xl p-8 bg-white border mt-32 mb-32">
     <div class="text-center">
-      <h2 class="text-3xl font-bold pb-2.5"><%= @user.name %>さんの解答</h2>
+      <h2 class="text-3xl font-bold pb-2.5"><%= current_user.name %>さんの解答</h2>
     </div>
     <div>
      <% if !@questions.empty? %>
@@ -19,7 +19,7 @@
      <% else %>
         <div>
           <p>問題を設定して下さい。</p>
-          <%= link_to '問題設定へ', questions_new_path %>
+          <%= link_to '問題設定へ', questions_new_path, class: 'underline text-blue-500' %>
         </div>
      <% end %>
     </div>
