+++ app/views/notification_settings/new.html.erb
@@ -26,7 +26,7 @@
       <% end %>
     </div>
     <div class="mt-2 text-xs">
-      <p>ラインのユーザー登録は<%= link_to 'こちら', line_users_new_path %></p>
+      <p>ラインのユーザー登録は<%= link_to 'こちら', line_login_api_new_path %></p>
     </div>
   </div>
 </div>
