+++ config/locales/views/ja.yml
@@ -36,6 +36,7 @@ ja:
     create:
       success: '正解しました！'
       fail: '間違えています'
+
